Displays 1st 2 Bytes
Learned how to use the Bochs Debugger.
Created the sectorDisplayFunctions_16.asm file for all of the code related to it.
Had issues as to why in the memory address the hex was 'FA' but I was displaying 'FD', the issue was I was calling the wrong loop subroutine.

Work Time 57 mins started at 10am